requirements.txt
torch==2.7.1
pytorch-lightning==2.5.2
numpy==2.3.1
pandas==2.3.0
scipy==1.16.0
PyWavelets==1.8.0
PyYAML==6.0.2
scikit-learn==1.7.0
tqdm==4.67.1
jupyter==1.1.1
matplotlib==3.10.3
seaborn==0.13.2